+++ btclib/ecc/dh.py
@@ -42,8 +42,7 @@ from __future__ import annotations
 from hashlib import sha256
 from math import ceil
 
-from btclib_secp256k1 import keys as libsecp256k1_keys
-
+from btclib._libsecp256k1 import keys as libsecp256k1_keys
 from btclib.alias import HashF, Point
 from btclib.curves import Curve, bytes_from_point, mult, secp256k1
 from btclib.curves.curve import _assert_valid_ec, _libsecp256k1_serves