+++ btclib/ecc/ssa.py
@@ -60,8 +60,7 @@ from hashlib import sha256
 from types import TracebackType
 from typing import overload
 
-from btclib_secp256k1 import ssa as libsecp256k1_ssa
-
+from btclib._libsecp256k1 import ssa as libsecp256k1_ssa
 from btclib.alias import BinaryData, HashF, Integer, JacPoint, Octets, Point
 from btclib.bip32 import BIP32Key
 from btclib.curves import Curve, PreparedPoint, secp256k1