+++ btclib/script/engine/tapscript.py
@@ -8,9 +8,8 @@ from __future__ import annotations
 
 from collections.abc import Mapping
 
-from btclib_secp256k1.ssa import verify as _libsecp256k1_ssa_verify
-
 from btclib import var_bytes
+from btclib._libsecp256k1 import ssa_verify as _libsecp256k1_ssa_verify
 from btclib.alias import ScriptList
 from btclib.curves import secp256k1
 from btclib.curves.curve import _libsecp256k1_serves